Move KeyOriginInfo to its own header file
